rename: loginController, loginView 파일명 수정
- 로그인 로직이 들어있을 것으로 오해할 것 같아서 수정함
