+++ src/navBar/index.ts
@@ -1,4 +1,4 @@
-import { showLoginModalOnLoad } from "./loginController";
+import { showLoginModalOnLoad } from "./loginModalController";
 
 export const initNavBar = () => {
   showLoginModalOnLoad();
